Prevent duplication of messages edited from other clients
Populus edits messages with events of type `m.reaction`, but other
clients, notably Element, use events of type `m.room.message`. This
filters out those events of this type with a `m.replace` relation.

It seems matrix-js-sdk is still smart enough to render the more
up-to-date version of the original message.