Add support for authenticated media
As per MSC3916.

In the service worker, we hook into the workbox caching routes in
order to add the auth token to the requests when they go out to fetch
from the network. The main thread retrieves the token from local
storage and sends it to the service worker, after it's activated.

We also change calls to functions provided by matrix-js-sdk, so that
they use authentication.