Revert service worker `/download`ing to legacy endpoint
It seems matrix-js-sdk still uses the legacy
`/_matrix/media/v3/download/` endpoint.

Element-web rewrites the URL in place in the service worker. See the
following for context:

https://github.com/element-hq/element-web/pull/27326#discussion_r1576303882

We won't do that and just hope that things get sorted out in
matrix-js-sdk eventually. This means populus remains broken with
homeservers that enforce authenticated media. See here for context and
how to disable it in current synapse, for the time being:

https://element-hq.github.io/synapse/v1.120/upgrade.html#authenticated-media-is-now-enforced-by-default