Filter out replacement and thread events
We filter replacement events and rely on matrix-js-sdk and server-side
aggregation

https://spec.matrix.org/v1.14/client-server-api/#server-side-aggregation-of-mreplace-relationships

to render the most current state of a message.

We also filter out thread messages for now. This potentially omits
messages / discussions but avoids messing up the timeline if and when
users open threads from other clients in our rooms. This is supposed
to be temporary until support for threads or a sane fallback is
implemented.

Closes #149