Fix fallback for replies
The function `stripFallbackPlain` was being passed a string, not an
array of lines. Changed to `stripFallbackPlainString` which expects a
string.

Not sure what the call to the `map` method was doing after
`getReplyPrefixPlain(content)`, which returns a string.

These issues were leading to type errors, preventing the timeline from
rendering properly in some scenarios. The problematic scenarios became
more common after clients stopped providing fallback for rich replies,
when it became no longer recommended by the spec after version 1.13:

https://spec.matrix.org/v1.14/client-server-api/#rich-replies