Rework authentication info exchange with service worker
Instead of preemptively sending it from the main thread, we make the
service worker ask for the authentication info on demand.  This seems
to improve reliability over restarts, when memory is wiped out.

We use two global variables in the service worker.  One to cache the
actual authentication info in memory and another to coalesce multiple
concurrent calls to `getAuthInfo` into a unique deliverable promise.

We also added some error handling, a timeout and a check to ensure the
token is only sent to the homeserver.