Let js-sdk handle the device ID
`loginRequest` does not update the client:

https://github.com/matrix-org/matrix-js-sdk/pull/4632

So, we also save the device ID to local storage on login.  This may
come in handy if we ever decide to support E2EE.

In `isResumable()`, sessions without `deviceId` in local storage are
still considered OK, since there is no reason to break current logged
in sessions over this.